# docs/swagger.yaml
basePath: /api
...
host: localhost:8080
info:
  contact: {}
  description: Live streaming platform API with chat
  title: GoLive API
  version: "1.0"
paths:
...
  /users/{userID}/follow:
    delete:
      description: Unfollow a user (Not implemented yet)
      parameters:
      - description: User ID to unfollow
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User unfollowed successfully
          schema:
            additionalProperties: true
            type: object
        "501":
          description: Not implemented
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Unfollow User
      tags:
      - Users
    post:
      description: Follow another user (Not implemented yet)
      parameters:
      - description: User ID to follow
        in: path
        name: userID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User followed successfully
          schema:
            additionalProperties: true
            type: object
        "501":
          description: Not implemented
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Follow User
      tags:
      - Users
  /users/me:
    delete:
      description: Delete the currently authenticated user's account (Not implemented
        yet)
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            additionalProperties: true
            type: object
        "501":
          description: Not implemented
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete Current User Account
      tags:
      - Users
    get:
      description: Get the profile information of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User profile information
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Current User Profile
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: Update the profile information of the currently authenticated user
        (Not implemented yet)
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            additionalProperties: true
            type: object
        "501":
          description: Not implemented
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update Current User Profile
      tags:
      - Users
  /users/me/followers:
    get:
      description: Get list of users that are following the current user (Not implemented
        yet)
      produces:
      - application/json
      responses:
        "200":
          description: List of followers
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "501":
          description: Not implemented
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Followers List
      tags:
      - Users
  /users/me/following:
    get:
      description: Get list of users that the current user is following (Not implemented
        yet)
      produces:
      - application/json
      responses:
        "200":
          description: List of followed users
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "501":
          description: Not implemented
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Following List
      tags:
      - Users
  /users/me/streams:
    get:
      description: Get all streams belonging to the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of user's streams
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get Current User's Streams
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"